+++ spec/dummy/db/migrate/20260904000000_add_type_to_test_contacts.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddTypeToTestContacts < ActiveRecord::Migration[7.1]
+  def change
+    add_column :contacts, :type, :string
+  end
+end